Add tiered store benchmark matrix
